@@ -1,17 +1,17 @@
 {
   "name": "mercadopago",
-  "version": "1.5.17",
+  "version": "2.0.1",
   "description": "Mercadopago SDK for Node.js",
-  "main": "index.js",
-  "files": [
-    "lib/"
-  ],
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "doc": "jsdoc -c jsdoc.json",
-    "lint": "NODE_ENV=test eslint lib/*.js",
-    "coverage": "NODE_ENV=test nyc ./node_modules/mocha/bin/_mocha ./test/*.*",
-    "coveralls": "npm run coverage -- --report lcovonly && cat ./coverage/lcov.info | coveralls",
-    "test": "NODE_ENV=test mocha ./test/*.*"
+    "build": "rm -rf dist && tsc --build tsconfig.production.json && tsc-alias -p tsconfig.production.json",
+    "test": "jest",
+    "test:e2e": "jest --config e2e/jest.config.ts",
+    "lint": "eslint 'src/**/*.{ts,tsx}' 'src/examples/**/*.{ts,tsx}' --cache",
+    "lint:fix": "eslint --fix --ext .ts src src/examples",
+    "exec": "node_modules/.bin/ts-node -r tsconfig-paths/register",
+    "prepare": "husky install"
   },
   "repository": {
     "type": "git",
@@ -26,30 +26,32 @@
     "sdk",
     "integration"
   ],
-  "author": "Ariel Rey <ariel.rey@mercadolibre.com>",
+  "author": "Mercado Pago (https://www.mercadopago.com/developers/en)",
   "bugs": {
     "url": "https://github.com/mercadopago/sdk-nodejs/issues"
   },
   "homepage": "https://github.com/mercadopago/sdk-nodejs#readme",
   "devDependencies": {
-    "chai": "3.5.0",
-    "chai-as-promised": "6.0.0",
-    "coveralls": "^3.1.0",
-    "eslint": ">=4.18.2",
-    "eslint-config-airbnb-base": "11.0.1",
-    "eslint-plugin-import": "2.2.0",
-    "jsdoc": "^4.0.2",
-    "mocha": "^9.1.3",
-    "mocha-lcov-reporter": "1.2.0",
-    "nyc": "^15.1.0",
-    "sinon": "1.17.7"
+    "@types/jest": "^29.5.3",
+    "@types/node": "^20.4.8",
+    "@types/node-fetch": "^2.6.4",
+    "@types/uuid": "^9.0.3",
+    "@typescript-eslint/eslint-plugin": "^6.3.0",
+    "@typescript-eslint/parser": "^6.3.0",
+    "eslint": "^8.46.0",
+    "husky": "^8.0.0",
+    "jest": "^29.6.2",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.1",
+    "tsc-alias": "^1.8.7",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.1.6"
   },
+  "files": [
+    "dist"
+  ],
   "dependencies": {
-    "ajv": "^6.12.3",
-    "bluebird": "3.4.7",
-    "dayjs": "^1.11.7",
-    "request": "^2.88.0",
-    "request-etag": "2.0.3",
-    "uuid": "3.0.1"
+    "node-fetch": "^2.6.12",
+    "uuid": "^9.0.0"
   }
 }
