@@ -1,85 +1,90 @@
-const mercadopago = new MercadoPago("TEST-2d36376d-c262-46bf-8480-3b2b27f81719", {
-    locale: "es-AR", 
-})
 
-
-document.getElementById("checkout-btn").addEventListener("click", function () {
-
-    // $('#checkout-btn').attr("disabled", true); ---------------------CUIDAD00000000000000000000000-----!!!!!!!!!!!!+65+56+w4e64965464985448
+  document.addEventListener("DOMContentLoaded", function() {
+    const mercadopago = new MercadoPago("TEST-2d36376d-c262-46bf-8480-3b2b27f81719", {
+      locale: "es-AR", 
+  })
   
-    const orderData = {
-      quantity: document.getElementById("quantity").value,
-      description: document.getElementById("product-description").innerHTML,
-      price: document.getElementById("unit-price").innerHTML
-    };
   
-    fetch("http://localhost:8080/create_preference", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(orderData),
-    })
-      .then(function (response) {
-        return response.json();
+  document.getElementById("checkout-btn").addEventListener("click", function () {
+  
+      // $('#checkout-btn').attr("disabled", true); ---------------------CUIDAD00000000000000000000000-----!!!!!!!!!!!!+65+56+w4e64965464985448
+    
+      const orderData = {
+        quantity: document.getElementById("quantity").value,
+        description: document.getElementById("product-description").innerHTML,
+        price: document.getElementById("unit-price").innerHTML
+      };
+    
+      fetch("http://localhost:8080/create_preference", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(orderData),
       })
-      .then(function (preference) {
-        createCheckoutButton(preference.id);
-  
-        $(".shopping-cart").fadeOut(500);
-        setTimeout(() => {
-          $(".container_payment").show(500).fadeIn();
-        }, 500);
-      })
-      .catch(function () {
-        alert("Unexpected error");
-        $('#checkout-btn').attr("disabled", false);
-      });
-  });
-  
-  function createCheckoutButton(preferenceId) {
-    // Initialize the checkout
-    const bricksBuilder = mercadopago.bricks();
-  
-    const renderComponent = async (bricksBuilder) => {
-      if (window.checkoutButton) window.checkoutButton.unmount();
-      await bricksBuilder.create(
-        'wallet',
-        'button-checkout', // class/id where the payment button will be displayed
-        {
-          initialization: {
-            preferenceId: preferenceId
-          },
-          callbacks: {
-            onError: (error) => console.error(error),
-            onReady: () => {}
+        .then(function (response) {
+          return response.json();
+        })
+        .then(function (preference) {
+          createCheckoutButton(preference.id);
+    
+          $(".shopping-cart").fadeOut(500);
+          setTimeout(() => {
+            $(".container_payment").show(500).fadeIn();
+          }, 500);
+        })
+        .catch(function () {
+          alert("Unexpected error");
+          $('#checkout-btn').attr("disabled", false);
+        });
+    });
+    
+    function createCheckoutButton(preferenceId) {
+      // Initialize the checkout
+      const bricksBuilder = mercadopago.bricks();
+    
+      const renderComponent = async (bricksBuilder) => {
+        if (window.checkoutButton) window.checkoutButton.unmount();
+        await bricksBuilder.create(
+          'wallet',
+          'button-checkout', // class/id where the payment button will be displayed
+          {
+            initialization: {
+              preferenceId: preferenceId
+            },
+            callbacks: {
+              onError: (error) => console.error(error),
+              onReady: () => {}
+            }
           }
-        }
-      );
-    };
-    window.checkoutButton =  renderComponent(bricksBuilder);
-  }
+        );
+      };
+      window.checkoutButton =  renderComponent(bricksBuilder);
+    }
+    
+    // Handle price update
+    function updatePrice() {
+      let quantity = document.getElementById("quantity").value;
+      let unitPrice = document.getElementById("unit-price").innerHTML;
+      let amount = parseInt(unitPrice) * parseInt(quantity);
+    
+      document.getElementById("cart-total").innerHTML = "$ " + amount;
+      document.getElementById("summary-price").innerHTML = "$ " + unitPrice;
+      document.getElementById("summary-quantity").innerHTML = quantity;
+      document.getElementById("summary-total").innerHTML = "$ " + amount;
+    }
+    
+    document.getElementById("quantity").addEventListener("change", updatePrice);
+    updatePrice();
+    
+    // Go back
+    document.getElementById("go-back").addEventListener("click", function () {
+      $(".container_payment").fadeOut(500);
+      setTimeout(() => {
+        $(".shopping-cart").show(500).fadeIn();
+      }, 500);
+      $('#checkout-btn').attr("disabled", false);
+    });
   
-  // Handle price update
-  function updatePrice() {
-    let quantity = document.getElementById("quantity").value;
-    let unitPrice = document.getElementById("unit-price").innerHTML;
-    let amount = parseInt(unitPrice) * parseInt(quantity);
-  
-    document.getElementById("cart-total").innerHTML = "$ " + amount;
-    document.getElementById("summary-price").innerHTML = "$ " + unitPrice;
-    document.getElementById("summary-quantity").innerHTML = quantity;
-    document.getElementById("summary-total").innerHTML = "$ " + amount;
-  }
-  
-  document.getElementById("quantity").addEventListener("change", updatePrice);
-  updatePrice();
-  
-  // Go back
-  document.getElementById("go-back").addEventListener("click", function () {
-    $(".container_payment").fadeOut(500);
-    setTimeout(() => {
-      $(".shopping-cart").show(500).fadeIn();
-    }, 500);
-    $('#checkout-btn').attr("disabled", false);
-  });
+});
+
